Add the beaker role to ci.yaml
Make the beaker.yaml a beaker role and add it to ci.yaml. I tend to
forget to run the extra role on a new beaker system. This way, all
servers listed in the host file under a beaker section get provisioned
the right way, before slave or master roles are applied.

Signed-off-by: Roman Mohr <rmohr@redhat.com>