Organize Prow deployment components
To ensure that different components are not being evicted at runtime, we
should have memory requests and limits set on every deployed container.
Currently, added default req and lim for all the containers and set it
to 510Mi. For components that are expected to consume more memory
(currently greenhouse and deck), we overwrite the defaults.
Also, added pod anti-affinity rules for components with more than one
replica. Also, move all the components to run on the ci zone.