Run Prow services on the CI zone (#97)
The infra zone consist of VMs that can migrate, thus we use
it for running Jenkins instances in order to achieve high availability
(we can't run replicas of Jenkins because they can't share the same
storage).

We prefer not to have other services on the nodes that runs the Jenkins
instances, in order to make it more stable.

Signed-off-by: gbenhaim <galbh2@gmail.com>