Tide for kubevirt/kubevirt (#134)
* Enable tide for kubevirt/kubevirt

Enable tide, tell tide that travis and coveralls are external required
tests.

* Mark our e2e test as required for tide

All e2e test are required, but don't run all of them until the PR is
approved.

The following tests will always run:

 * pull-kubevirt-e2e-k8s-1.13.3
 * pull-kubevirt-e2e-k8s-multus-1.13.3
 * pull-kubevirt-e2e-os-3.11.0-crio

The following tests will only be run when a PR is in the merge pool:

 * pull-kubevirt-e2e-k8s-1.11.0
 * pull-kubevirt-e2e-k8s-genie-1.11.1
 * pull-kubevirt-e2e-os-3.11.0-multus
 * pull-kubevirt-e2e-windows2016

Tests which by default only run in the merge-pool can still be triggered
via `/test <name>` or `/test all` and will be rerun if `/retest` is done
afterwards.