Allow jobs which run always to run more often
Now we only run a subset of jobs when PRs are opened. They can now run
more often, since the other tests don't run by default that often
anymore. They only run when the PR is in the mergepool or if they are
manually triggered. This should help to bring PR faster into the merge
pool.