Virt Blocks update (#172)
* virtblocks: Use custom container image

This image is built from

  https://github.com/virtblocks/virtblocks/blob/master/extras/Dockerfile

and it's basically the same as kubevirt/golang-libtool:1.12.5-r1,
except it also contains the Rust compiler and tools.

* virtblocks: Don't set VIRTBLOCKS_LANGUAGE

As of today, setting the variable causes the Makefile machinery
to only build and test the Go code, but now that we're using a
container image that includes a Rust compiler as well we can
extend testing to that part of the repository too.

* virtblocks: Run test suite and example programs

We have some test cases now, with more hopefully on the way, and
a convenient way to run example programs too, so let's do that
when pull requests come in.