Sort report data descending by severity
Sort tests descending by severity, so that worst (red) test cells are
shown at the top of the report.
Furthermore add --preview flag to not meddle with existing reports, i.e.
old reports are not overwritten any more if flag is used, if flag is used
reports are written to .../preview subdirectory instead.