Set TARGET_COMMIT to base sha (#397)
Prow checks out the base branch and merges the pr branch
on top of it, so the last merge commit is ofcourse the
current HEAD. This leads to "no files changed" when trying to
fetch changed files after last merge commit (which is done
by default in the script.
Therefore we explicitly set TARGET_COMMIT to the base branch
sha.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>