sriov: run lane on each PR
SR-IOV deployments are broken more often than not, and that is
easily explained since it does not vote in CI, and does not even
run.

This is a step in that direction; run it, but don't have it voting.
We intend to actively debug the sriov lane, and nurture it back to
a healthy state, then have it voting in CI.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>