Add automation to configure remote cluster (#469)
* Add remote K8s cluster to execute Prow jobs

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>

* Mount kubeconfig to Prow components

In order to execute jobs on a remote cluster, a kubeconfig has to be
mounted to Prow's core components.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>

* Move prow jobs secrets to a dedicated role

As a preperation for multi-cluster deployment, moved the secrets that
are required to run Prow jobs into a dedicated reusable role.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>

* Update README file

Since we now feature a multi-cluster deployment, I've updated the docs
to the new flow.
Also updated the `how to share secrets` section to our new method.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>

* Move shared components to a dedicated role

Since we have multiple execution clusters for Prow jobs, we have a
shared set of components that we need to deploy in order to configure
the namespaces where the Prow jobs will be running in.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>