Change weekly flakefinder report to run as cron
As of latest changes to flakefinder the report data coverage interval
is normalized against start of resp. end of day to avoid overlapping PRs
in two adjacent reports. Now the flakefinder weekly reports are run
somewhere during the day using `interval` which may lead to them running
late in the day and therefore not being available.

We use cron and let them run somewhere after midnight (also adding some
variance) to have them available at the start of the day for when the
build officer starts working.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>