Add an optional lane with a newer version of base image.
The CDI tests now require libdevmapper.so to run. The base image all the
lanes are running from does not include it, but a latest version does.
This commit adds an optional lane to verify that a/ the tests run
correctly and b/ the CI works with the updated image.

Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>