labels: Rename topic/* to sig/* (#580)
The label plugin supports labeling using a command only specific types
and "topic" is not one of them.

"sig" is a supported type and is equivalent in meaning to the "topic"
one.
With this change, one can label a PR or issue using the
`/sig <name>` comment.
The name can be any of the following: network, storage, virtualization
and security.

Signed-off-by: Edward Haas <edwardh@redhat.com>