Unblock status reconciler (#600)
Making `pull-kubevirt-generate` required on master, but not on release
branche confuses status-reconciler. As a quick solution unblock it by
removing the unused jobs from the release-branches.

Status-reconciler sees the job the once as optional and once as required
in the current job list and fails to distinguish them. When it compares
the last known state with the current state it can't distinguish them
and thinks that the second optional job for the release branch is now
required and retriggers the jobs in an endless loop.

Signed-off-by: Roman Mohr <rmohr@redhat.com>