Add kubernetes_crud role
this role a is a generic interface between ansible and batch deployments
of kubernetes manifests.
It allows to write steps in a declarative way avoiding duplication of
common parameters such as kubeconfig or context.
This role will be used by the upcoming rewrite of role to deploy prow
and its dependencies.
A CI test has been added that creates a kubevirtci cluster inside a
container. This job will be the base for all other deployment tests.

Signed-off-by: Gabriele Cerami <gcerami@redhat.com>