Mitigate docker hub push fails on EOF with RestartPolicy=OnFailure
In recent times the push to docker hub started to fail with just a short
error like this:

Error pushing image to ...: unable to push image to ...: Put
https://index.docker.io/v2/kubevirtnightlybuilds/...: EOF

Locally this could be reproduced once, on second and third try it
succeeded. Therefore we change the container restartPolicy to OnFailure
to retry the push.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>