Close build watcher issues after 14 days (#771)
In the current state, we often forget to close flakey issue report after
the problem gets resolved. Due to this we end up with irrelevant Issues
rotting around for months.

With this change, we will close all issues triaged by the build watcher
after a week of inactivity. If an issue reoccurs, it should be mentioned
on the Issue.

Signed-off-by: Petr Horáček <phoracek@redhat.com>