sriov cni, Allocate only one sriov PF for the target container
In order to be able to run multi sriov clusters on CI sriov node, at the same time,
each Job should get only one PF (per CNI call), and not all the available PFs.

Update this folder to contain only the updated CNI.

This CNI assume multus is already deployed.
It can run side by side with the previous CNI, as it
has a different name.

Signed-off-by: Or Shoval <oshoval@redhat.com>