sriov CNI, Allocate only one sriov PF
In order to allow running multi sriov clusters simultaneously,
allocate only one sriov PF per CNI call.

In case no PFs exist, exit with an error, so the job won't
start without PF.
It may occur in case the container was forcely deleted,
and then it takes a few minutes for the PF
to reappear in /sys/class/net.
The CNI will exit, delete the old NS, and a new container will
be created and retrigger CNI that would try again, until
the PF reappears.

Signed-off-by: Or Shoval <oshoval@redhat.com>