sriov, Change CNI annotation style
With this updated annotation style,
one can call the same CNI twice by duplicating the annotation,
(must be in the same annotation line) and changing only the interface name.

It will allows to call the updated CNI which allocate only one PF
twice, in order to allocate 2 PFs for migration.

Signed-off-by: Or Shoval <oshoval@redhat.com>