Remove version dependent bazel handling
The version checks are not relevant for us and they get confused if the
bazel version is not set. Since we want to detect the bazel version
based on the project via bazelisk, this is right now irrelevant and can
be re-added in an appropriate form later.

This fixes issues like this:

```
ERROR: --experimental_remote_spawn_cache :: Unrecognized option: --experimental_remote_spawn_cache
```

when building with bazel directly in the image and no BAZEL_VERSION was
explicitly set.

Signed-off-by: Roman Mohr <rmohr@redhat.com>