sriov CNI, Allocate only one sriov PF (#884)
* sriov CNI, Remove unneeded conf file

The NAD configuration that is used is the
NetworkAttachmentDefinition, not the conf file one.
Hence remove it.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* sriov CNI, Remove unneeded code

veth code is not used, nor needed.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* sriov CNI, Allocate only one sriov PF

In order to allow running multi sriov clusters simultaneously,
allocate only one sriov PF per CNI call.

In case no PFs exist, exit with an error, so the job won't
start without PF.
It may occur in case the container was forcely deleted,
and then it takes a few minutes for the PF
to reappear in /sys/class/net.
The CNI will exit, delete the old NS, and a new container will
be created and retrigger CNI that would try again, until
the PF reappears.

Signed-off-by: Or Shoval <oshoval@redhat.com>