Make the dependency mirror tool more fault  tolerant (#970)
* Don't upload error body on failed http gets

If we don't get a 2XX response, we should not updload the returned body.

Signed-off-by: Roman Mohr <rmohr@redhat.com>

* Verify shasum of rpm source

Signed-off-by: Roman Mohr <rmohr@redhat.com>

* Add a flag to continue uploads on error

The exit code of the command will still be non-zero in case that an
artifact could not be uploaded.

Signed-off-by: Roman Mohr <rmohr@redhat.com>

* Add a command to verify all mirrored artifacts

Signed-off-by: Roman Mohr <rmohr@redhat.com>

* Use authenticated client for verification if not a dry-run

Anonymous users get results from cache which can be outdated for some
time which can confuse the validation.

Signed-off-by: Roman Mohr <rmohr@redhat.com>