kubevirt-presubmits: Add sig-network k8s 1.19 job
The purpose of this job is to allow running a more
focused group of tests, in isolation from the other tests,
owned by the sig-network.

This lane would always run, but won't be a gating.
It will run only the tests that are marked with sig-network,
including MacVtap and Multus (which require CNAO).

Signed-off-by: Or Shoval <oshoval@redhat.com>