Set the bazel output base directory to a well known location (#1020)
Ensure that we know which directory bazel will use as a cache
directories, independent of the user used on the prow job.

Further mounting an emptyDir at this well known location if jobs decide
to run bazel unnested.

This way bazel will not write massively into the container overlay,
which can fill up the container space pretty fast and which can also
perform bad.

Signed-off-by: Roman Mohr <rmohr@redhat.com>