Skip job data outside report window for periodics
We never want to include job data that is outside the time window of
the report.
This differs from the PR based data retrieval where we
want to include all job data from a PR, even if it is before the time
window, to have the complete picture of all failures.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>