Add periodic job, code and simple test for kubevirt-job-copier (#1511)
* Add periodic job, code and simple test for kubevirt-job-copier

kubevirt-job-copier creates new periodic and presubmit jobs in case a
new kubernetes provider gets released. It does that by copying the
existing jobs that use the currently latest kubevirt provider.

Periodics cron expressions are advanced to spread the load, presubmits
are created as optional and always_run: false. This way they are not
failing when the new provider has not landed in the kubevirt codebase
yet, but can be triggered manually when it has landed.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Address review comments

Minor refactoring: replace log parameter in favor of function

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix case for two patch releases out

Signed-off-by: Daniel Hiller <dhiller@redhat.com>