Add check that lgtm label is missing before bumping kubevirtci (#1528)
* Add check that lgtm label is missing before bumping kubevirtci

Adds some code that checks whether certain labels are missing and fails
with non zero exit code if they are not.

Changes the bump-kubevirtci job to use that code to ensure lgtm
label is missing on PR.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add simple description of labels-checker to README

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update builder bazel version for running labels-checker

Also add missing change directory.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Address review comments

Fix several comments, also add approved as another label to check for on
the job definition.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>