Create pr-creator image (#1536)
* Create pr-creator image

Adds an image as base for jobs using git-pr.sh. This is intended to
replace the kubevirt builder image where that image is not required to
be used.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add scripts git-pr and git-askpass

We add two scripts from hack dir, git-askpass.sh and git-pr.sh to the
path of the image and also set the env var GIT_ASKPASS so we do not have
to repeat this on all jobs.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add labels-checker

Add labels-checker to use it directly in contrast to having to use
bazel to run it.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>