Add a GOPROXY fallback if the main proxy rejects us
Add a second proxy and ensure that go will download from the next
specified source in case of any failure.

The old value was https://proxy.golang.org,direct, the new value is
https://proxy.golang.org|https://goproxy.io|direct.

This fixes an issue where the jobs sometimes gets a GOAWAY which leads
to an immediate error, since direct download is only considered if a 404
or a 410 error code is returned.

Signed-off-by: Roman Mohr <rmohr@redhat.com>