Add merging of several junit artifacts for accurate reports
Due to the several artifacts on a build we get random results in the
report. Therefore we fetch all artifacts and merge those into one. We
use the shortcut of ignoring duplicated test entries and adding a note
on the suite instead.

TODO: this needs to get handled in the report, so we can directly see
that we need to fix this u/s

Signed-off-by: Daniel Hiller <dhiller@redhat.com>