Add flake-report-creator for jenkins (#1897)
* initial jenkins report creation

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Parallelization and refactoring

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add links to test lanes and builds

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Limit number of connections for http.Client

Also, instead of fetching all jobs, just get the job names to avoid
querying for every job.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add merging of several junit artifacts for accurate reports

Due to the several artifacts on a build we get random results in the
report. Therefore we fetch all artifacts and merge those into one. We
use the shortcut of ignoring duplicated test entries and adding a note
on the suite instead.

TODO: this needs to get handled in the report, so we can directly see
that we need to fix this u/s

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Reorganize the jenkins tool to sit below flake-report-creator

Also we reorganize the code to use cobra command library and
generalize handling of output files.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Sort Jobs by BuildNumber

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Improve logging

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update bazel

Add bazel stuff for image build, add push to Makefile

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add ./hack/flake-report-creator.sh

Add script to execute flake-report-creator using docker.

Example:

```
./hack/flake-report-creator.sh jenkins
```

Update examples in README.md to cover the script usage.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix tests

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix concurrent data fetching

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add test for merging

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix error in artifact retrieval, add test

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update image in script

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Replace cnv versions with a more generic pattern

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add insecureSkipVerify

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update script to use new image

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix README

Signed-off-by: Daniel Hiller <dhiller@redhat.com>