Ensure that vgpu jobs are not scheduled to the same nodes (#1955)
We reconfigure the whole GPU when executing vgpu jobs. During that time
no other vgpu job should get scheduled to the node, so that they are not
in the way of each other.

Add a pod anti affinity which ensures that pods labeled with `vgpu:
true` are not scheduled on the same nodes at the same time.

Signed-off-by: Roman Mohr <rmohr@redhat.com>