sig-performance: Increase KUBEVIRT_MEMORY_SIZE to 10G (#2064)
* sig-performance: Increase KUBEVIRT_MEMORY_SIZE to 10G

Thanks to 9f66270 we can now see that the sig-performance job was timing
about because of 12 unschedulable VMIs due to memory pressure on the
hosts.

This change attempts to resolve this entirely by bumping the overall
memory in the env by 4G (1G for each host) hopefully allowing these
final 12 VMIs to schedule.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

* Revert "sig-performance: Increase timeout to 1h30m0s"

This reverts commit 9f66270b4cc4a3ad86dce1da15de312d35dc5b89.

This shouldn't be required anymore with enough memory available in the
environment to schedule all of the VMIs.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>