Sort test results by severity
If tests have the same severity it can happen that tests with lesser
failures but more occurrences appear before tests with much more
failures. Example:

https://storage.googleapis.com/kubevirt-prow/reports/flakefinder/kubevirt/kubevirt/flakefinder-2022-04-11-168h.html#row42

Therefore instead of number of occurrences we add a weight that we
calculate from the number of failed tests divided by the ratio of failed
to succeeded tests. This will make a test with many failures appear
earlier in the report.

Example report preview:
https://storage.googleapis.com/kubevirt-prow/reports/flakefinder/preview/kubevirt/kubevirt/flakefinder-2022-05-01-168h.html

Comparable to original report:
https://storage.googleapis.com/kubevirt-prow/reports/flakefinder/kubevirt/kubevirt/flakefinder-2022-05-01-168h.html

This shows that i.e. the SRIOV failures that are happening a lot are
much higher now as they should be.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>