sriov: Run cluster-down only once in trap (#2074)
* sriov: Run cluster-down only once in trap

Since the trap had ERR as well as EXIT,
it would run the trap twice in case an error happened,
one for the ERR and one for the EXIT.
Having it on EXIT is enough because it happens
either on good flow and error flow.

SIGTERM is not needed, because EXIT will happen
anyway as well.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* sriov: Remove kind-1.19-sriov job

kind-1.22-sriov replaced this job

Signed-off-by: Or Shoval <oshoval@redhat.com>