flakefinder: fix timeouts for jenkins build retrieval (#2081)
* Replace explicit loading of latest build with number

As we only need the number of the latest build in the function, instead
of calling getLatestBuild, we directly use the latest build number from
the raw data as an argument to the function.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add retry to GetBuild

GetBuild suffers from being not robust to 504 errors when a gateway is
present, thus we add a retry for 504s.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>