Create a serviceaccount and Roles for debugging prow
We create a serviceaccount with limited access for looking at prow jobs
and pods on the cluster in general.

Based on this service account we could create a script that would fetch
the current tokens from the used KUBECONFIG and create a new config with
a newly attached service account token.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>