Add job-config-validator (#2058)
* Add job-config-validator

This adds the job-config validator that fetches a set of job config
files from a path, loads a config map and checks whether all job config
files are present in the map.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add support for GZIP configMap entries

If GZIP is used for ConfigMapSpec then the entries are put inside of the
BinaryData, thus we need to check those too.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add indirect dependency

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add periodic job checking job configs for prow

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Run goimports, remove unused dry-run

Signed-off-by: Daniel Hiller <dhiller@redhat.com>