Add prow debug serviceaccount and KUBECONFIG generation script (#2105)
* Create a serviceaccount and Roles for debugging prow

We create a serviceaccount with limited access for looking at prow jobs
and pods on the cluster in general.

Based on this service account we could create a script that would fetch
the current tokens from the used KUBECONFIG and create a new config with
a newly attached service account token.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add script to generate kubeconfig for user

We generate a new KUBECONFIG based on the certificate data from the
currently used KUBECONFIG, combined with a newly generated secret that
is attached to the serviceaccount prow-deploy.

Defaults of the KUBECONFIG all point to the relevant namespace.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>