Add priority class for windows e2e jobs (#2162)
Currently there is only a maximum of 1 windows e2e job allowed run at a
time.

This can lead to build up of windows job queues.

Due the restriction of the 1 job at a time - there should always be one
job running if there is a queue. This is not the case at the moment as
other regular e2e jobs can get scheduled ahead of them.

Adding a priority class for the windows jobs should ensure that the
windows job will get scheduled quicker when there is a queue.

Included cleanup of the windows priority class in case updates need to
be made.

Signed-off-by: Brian Carey <bcarey@redhat.com>