Update golang image to use go 1.17.8 (#2177)
* Update golang image to use go 1.17.8

The kubevirt builder image uses go version 1.17.8[1] - the golang image
used for testing should default to the same.

[1] https://github.com/kubevirt/kubevirt/blob/fadd752f0392a320483524aae753fe53930a6ba6/hack/builder/Dockerfile#L7

Signed-off-by: Brian Carey <bcarey@redhat.com>

* Include build of golang image in build-bootstrap-image presubmit

The image build file for the golang image requires a locally built
bootstrap image so the build-golang-image job fails as this is not
available.

These images get published together as part of the publish-bootstrap-image[1] job so it makes sense to test them together in the same presubmit.

[1] https://github.com/kubevirt/project-infra/blob/600cd512cbb691637c356cf192263ebd2c3f0462/github/ci/prow-deploy/files/jobs/kubevirt/project-infra/project-infra-postsubmits.yaml#L108

Signed-off-by: Brian Carey <bcarey@redhat.com>