Remove -gcs-workers flag from crier deployment
The automated prow bump PR was failing the presubmit deploy test due to crier
failing to start successfully.

Crier was failing to start due to an undefined flag being specified.
Removing the -gcs-workers flag allows crier to start successfully.

Also small update to prow-deploy README to include the correct path to
the kubeconfig.

Signed-off-by: Brian Carey <bcarey@redhat.com>