Skip large e2e tests when only docs are updated
Currently the full set of e2e presubmits run for simple doc updates.
This can be quite costly in CI time and should not be a requirement for
documentation updates.

Using the skip-if-only-changed configuration[1] to skip e2e jobs when there
is only updates to files under docs/ or only to .md files.

[1] https://github.com/kubernetes/test-infra/blob/master/prow/jobs.md#triggering-jobs-based-on-changes

Signed-off-by: Brian Carey <bcarey@redhat.com>