Update flake-report-creator to show failuresForJobs in report and json (#2190)
* Change failuresForJobs key to string

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add number of failures to builds

Add a popup that shows the overview of failures for each build.
Add the total number of failures per build to the test popup.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Extend json with failuresForJobs

We extend the created json to be a map with two keys,
`data` and `failuresForJobs`, where the former holds the original data
and the latter holds the aggregated number of failures per job run.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update flake-report-creator script to latest image

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

Signed-off-by: Daniel Hiller <dhiller@redhat.com>