Move build of multiarch bootstrap to weekly periodic (#2264)
Since the addition of the arm64 bootstrap image builds the presubmit and
postsubmit prowjobs take between 4 to 5 hours[1]. This can cause
issues when trying to merge updates to the bootstrap image.

This change moves the publishing of the arm64 images to a weekly
periodic job and returns the bootstrap presubmit and postsubmit jobs to
just build the x64_86 variant of the images.

[1] https://github.com/kubevirt/project-infra/issues/2252

Signed-off-by: Brian Carey <bcarey@redhat.com>

Signed-off-by: Brian Carey <bcarey@redhat.com>