Add netavark configuration for podman bootstrap image
Podman changed the network backend to netavark in v4.0 [1]. The old bridge
cni configuration no longer works with the netavark backend.

This change adds the default netavark configuration with IPv6 enabled
which fixes an issue with IPv6 outbound traffic in the conformance
tests [2]. It also removes the old cni configuration.

[1] https://github.com/containers/podman/releases/tag/v4.0.0
[2] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_project-infra/2273/rehearsal-check-provision-k8s-1.23/1561980262324113408

Signed-off-by: Brian Carey <bcarey@redhat.com>