Switch non-root sriov presubmit to optional
The non-root sriov presubmit lane has been failing consistently.[1]
This issue has not been seen in the periodic sriov lane which runs
rootful. The issue looks to be non-root specific.

The sriov PFs are not available in the prow pod.

Updating the non-root sriov presubmit to optional and the rootful sriov
presubmit to required. This allows us to keep testing sriov without
blocking unrelated PRs while investigation into non-root issue is
ongoing.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/8627/pull-kubevirt-e2e-kind-1.22-sriov-nonroot/1583370129192783872

Signed-off-by: Brian Carey <bcarey@redhat.com>