Increase wait time for podman socket to be ready
Timeouts waiting for the podman socket to be ready have been seen[1] when
there is high load on the cluster. Increasing the wait time should help
with this.

This change also removes the enabling of ip6table_nat as this looks to
be not needed and just causes an error when enabling ipv6.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/8605/pull-kubevirt-e2e-k8s-1.24-sig-network-nonroot/1581994128529428480

Signed-off-by: Brian Carey <bcarey@redhat.com>