Use correct provider name for operator lane (#2413)
Since we are using sig-operator in k8s <= 1.24 already, let's update the
lane names to reflect reality for 1.25.

The update of the sig name for the automation will be done in a separate
PR.

NOTE: we are NOT updating the lanes for 1.24 <= to avoid losing
information in the testgrid dashboard!

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

Signed-off-by: Daniel Hiller <dhiller@redhat.com>