We don't update the state if the retry is blocked
In case the interval for retry blocked has not yet expired
we do not update the circuit state with a new interval. This is
in order to not extend the interval if requests from a later point
in time than the initial opening request return with the same
error state.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>