Add circuit breaker to flake-report-creator (#2414)
* Add tests for handling 404 and 504

To make this testable, create new interface BuildDataGetter for
dependency injection with default implementation. Extract method that
uses the interface while preserving the original.

Also add pending test for parallel execution that already proves that
the service is called many more times than wanted.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add circuit breaker implementation

Also add another test that ensures that each underlying service call is
done at most two times from each thread

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Extract circuitbreaker to own package

Add unit tests

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add function that decides whether to open circuit

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Use sync.RWMutex, rephrase occurred to blockedUntil

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Correct the test descriptions and expectations

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* We don't update the state if the retry is blocked

In case the interval for retry blocked has not yet expired
we do not update the circuit state with a new interval. This is
in order to not extend the interval if requests from a later point
in time than the initial opening request return with the same
error state.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Remove redundant check

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Move the lock to before the isRetryBlocked call

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Change default retry to Fixed with jitter

We don't want to use default retry.BackOffDelay since it would multiply
the wait time, which initially starts with 3 mins, instead we replace it
with retry.FixedDelay with RandomDelay added.

Also we add logging to circuitbreaker so log readers are able to
understand what is going on.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update image in script

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

Signed-off-by: Daniel Hiller <dhiller@redhat.com>