Add files to build shared-images-controller image
This image is based off of the bootstrap image as the podman go bindings
require a running podman system socket to connect to.

Currently this image will just periodically download the latest kubevirtci node
images that are required by kubevirt/kubevirt main to the shared images
store. It also removes any old images that do not match this latest
kubevirtci tag.

The storage.conf is based off of the default config just with the
read/write location of container storage updated to match the path of
the shared images location.

Signed-off-by: Brian Carey <bcarey@redhat.com>