Update CDI image postsubmits to use podman
There was an issue with pushing CDI images when the prowjobs were
running with the podman in container setup. This issue was due to bazel
push expecting the registry auth information to exist under the docker location.

This change depends on a PR[1] in CDI which makes the required registry
auth config available in the builder container that pushes the images

This reverts commit d751dca8ff84463a950a7c78486f5f17aba92502.

[1] https://github.com/kubevirt/containerized-data-importer/pull/2479

Signed-off-by: Brian Carey <bcarey@redhat.com>