Add daemonset for shared images to prow-workloads cluster
This daemonset is responsible for maintaining the shared image store on
each node in the workloads cluster. It will download the latest
kubevirtci node images and remove any old images that are no longer
required. The e2e jobs will mount this directory as a read only shared
image store.

Signed-off-by: Brian Carey <bcarey@redhat.com>