Improve flakefinder test name layout (#2549)
* Externalize the HTML templates

We externalize the HTML template into their own files in order to
support IDE plugins for go templates that don't work with inline
templates. Along the way we create validation tests and fix the errors
in the markup.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix report template

Some corrections to the html have been applied, since now a validator
can show them.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add test attributes

Adds a sorted set of test attributes to the flakefinder.Params struct
that can be iterated over to separately show each tag.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Show test attributes as colored labels, strip test name

We integrate the test attributes as a set of colored labels
that are show on top of the test name cell. Also we now print only the
bare test name without the labels.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Improve visibility of name and value for attributes, copy button

This separates the name of the attribute and the value visually, also it
adds a button that copies the full test name into the clipboard.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix layout for labels, add some labels

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Fix formatting in source files

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update flakefinder images in jobs

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Update bazel build file for flake-report-creator

Build file was missimg embed information and some test files.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>