Port switch to centos9 for 1.26 to release 0.59 (#2611)
Ports the lane split for operator tests to 0.59 since we want to test
all configuration of operator with PSA enabled, while the upgrade can
not work, as 0.58 is not PSA ready.

Sets the gating lanes to use CentOS Stream 9 based providers and enables
PSA on all 1.26 lanes, except the operator-upgrade lane.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>