Add csv output to kubevirt get (#2617)
* Add csv output to kubevirt get

Adds flag --output-format where allowed values are html and csv,
according to the flag value an output file in that format is generated.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* Add read of multiple files for presubmits

We add the support for reading multiple files and merging the presubmit
configs into one report.

We also modify the postsubmit job configuration to pick up the presubmit
config for release-0.59.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>